requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==2.2.15
django-mailbox==4.8.1
djangorestframework==3.11.0
idna==2.10
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
